# customer/customer-api-spec/src/main/resources/openapi/openapi.yml
openapi: 3.0.1
info:
  title: Customer API
  version: 1.0.0
  contact:
    name: Manuel Mašlonka

servers:
  - url: http://localhost:8080/api/v1

paths:
  /customers:
    post:
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Customer created
        '400':
          description: Invalid request
        '409':
          description: Customer already exists

  /customers/add-account:
    post:
      summary: Add account to customer
      operationId: addAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId:
                  type: integer
                  format: int64
                accountId:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Successful response
        '400':
          description: Invalid request
        '404':
          description: Customer or account not found

  /customers/{Id}:
    get:
      summary: Get customer details
      operationId: getCustomer
      parameters:
        - name: Id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found

components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        accountIds:
          type: array
          items:
            type: integer
            format: int64

    CustomerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string